avoid referencing same article in similar articles, fix console warning
